Created new CSS and Register mockup
Created a separate stylesheet called stylesAdditions.css for any new
styles needed to be added. Created a register mockup page and  added
the text from Tuong.

Added the new timer icon to the box with the important information.
that Perry created.

Working on adding the collapse function to the virtual, full and
partial and the 1 2 3 icons for virtual, full and partial registration.

Conflicts:
	.DS_Store
	css/screen.css
	img/canheiticon-buytickets-120x100.png
	your-stay/index.html